pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Study project on django - 'Task manager'"
authors = ["Dzigr <dmbrag@gmail.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
python-dotenv = "^1.0.0"
dj-database-url = "^2.0.0"
psycopg2-binary = "^2.9.6"
pip = "^23.1.2"
gunicorn = "^20.1.0"
django-bootstrap5 = "^23.3"
whitenoise = "^6.5.0"
django-crispy-forms = "^2.0"
crispy-bootstrap5 = "^0.7"
django-filter = "^23.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
coverage = "^7.2.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
